refactor(design-tokens): move color plugins to @rafters/color-utils (#1244)
Color-domain plugins (scale, state, contrast) move to @rafters/color-utils/plugins.
Scale position constants move to @rafters/shared. Invert plugin deleted (unused).
design-tokens now has zero imports from color-utils in src/.
