+++ packages/color-utils/src/plugins/scale.ts
@@ -6,10 +6,9 @@
  * The resolveColorReference step in applyComputed converts it to a CSS string.
  */
 
-import { ColorReferenceSchema, ColorValueSchema } from '@rafters/shared';
+import { definePlugin } from '@rafters/design-tokens';
+import { ColorReferenceSchema, ColorValueSchema, INDEX_TO_POSITION } from '@rafters/shared';
 import { z } from 'zod';
-import { definePlugin } from '../plugins';
-import { INDEX_TO_POSITION } from '../scale-positions';
 
 const ScalePositionInputSchema = z.object({
   familyColorValue: ColorValueSchema,
