+++ packages/color-utils/src/plugins/state.ts
@@ -6,10 +6,9 @@
  * from the base token's actual position.
  */
 
-import { ColorReferenceSchema, ColorValueSchema } from '@rafters/shared';
+import { definePlugin } from '@rafters/design-tokens';
+import { ColorReferenceSchema, ColorValueSchema, INDEX_TO_POSITION } from '@rafters/shared';
 import { z } from 'zod';
-import { definePlugin } from '../plugins';
-import { INDEX_TO_POSITION } from '../scale-positions';
 
 const StateInputSchema = z.object({
   familyColorValue: ColorValueSchema,
