+++ packages/design-tokens/test/semantic-cascade.test.ts
@@ -1,6 +1,5 @@
 /**
  * Tests for the semantic token cascade system:
- * - scale-positions shared utilities (findBestWcagPair, findDarkCounterpartIndex)
  * - deriveGenerationRule pattern matching
  * - applyComputed updating dependsOn on cascade
  * - Repeated cascade stability
@@ -9,145 +8,15 @@
  * Plugin execution tests are in plugins.test.ts.
  */
 
-import type { ColorReference, ColorValue, Token } from '@rafters/shared';
+import type { ColorReference, Token } from '@rafters/shared';
 import { describe, expect, it } from 'vitest';
 import { generateBaseSystem } from '../src/generators';
 import { generateSemanticTokens } from '../src/generators/semantic';
 import type { ResolvedSystemConfig } from '../src/generators/types';
 import { TokenRegistry } from '../src/registry';
-import {
-  findBestWcagPair,
-  findDarkCounterpartIndex,
-  MIN_WCAG_PAIR_DISTANCE,
-} from '../src/scale-positions';
 
 // Minimal config for generateSemanticTokens
 const minimalConfig = {} as ResolvedSystemConfig;
-
-// Helper: build a ColorValue with known WCAG pairs
-function makeColorValueWithPairs(aaaPairs: number[][], aaPairs: number[][] = []): ColorValue {
-  const scale = Array.from({ length: 11 }, (_, i) => ({
-    l: 0.95 - i * 0.09,
-    c: 0.1,
-    h: 240,
-    alpha: 1,
-  }));
-
-  return {
-    name: 'test-color',
-    scale,
-    accessibility: {
-      wcagAAA: { normal: aaaPairs, large: [] },
-      wcagAA: { normal: aaPairs, large: [] },
-      onWhite: { wcagAA: true, wcagAAA: false, contrastRatio: 5 },
-      onBlack: { wcagAA: true, wcagAAA: false, contrastRatio: 5 },
-      apca: { onWhite: 60, onBlack: -60 },
-    },
-  };
-}
-
-describe('findBestWcagPair', () => {
-  it('finds the partner with greatest distance on the higher side', () => {
-    const pairs = [
-      [0, 5],
-      [0, 8],
-      [0, 10],
-    ];
-    expect(findBestWcagPair(0, pairs, true)).toBe(10);
-  });
-
-  it('finds the partner with greatest distance on the lower side', () => {
-    const pairs = [
-      [10, 5],
-      [10, 2],
-      [10, 0],
-    ];
-    expect(findBestWcagPair(10, pairs, false)).toBe(0);
-  });
-
-  it('skips partners on the wrong side', () => {
-    const pairs = [
-      [5, 3],
-      [5, 2],
-      [5, 1],
-    ];
-    // Want higher, but all partners are lower
-    expect(findBestWcagPair(5, pairs, true)).toBeUndefined();
-  });
-
-  it('handles both orderings in pairs', () => {
-    // [a, b] where b === sourceIndex
-    const pairs = [[8, 2]];
-    expect(findBestWcagPair(2, pairs, true)).toBe(8);
-  });
-
-  it('returns undefined for empty pairs', () => {
-    expect(findBestWcagPair(5, [], true)).toBeUndefined();
-  });
-
-  it('skips malformed pairs', () => {
-    const pairs = [[], [1], [undefined as unknown as number, 5]];
-    expect(findBestWcagPair(5, pairs, true)).toBeUndefined();
-  });
-});
-
-describe('findDarkCounterpartIndex', () => {
-  it('uses AAA pairs first', () => {
-    const cv = makeColorValueWithPairs(
-      [[0, 10]], // AAA: 0 pairs with 10
-      [[0, 6]], // AA: 0 pairs with 6
-    );
-    expect(findDarkCounterpartIndex(0, cv)).toBe(10);
-  });
-
-  it('falls back to AA when AAA is too close', () => {
-    const cv = makeColorValueWithPairs(
-      [[5, 7]], // AAA: distance 2, below MIN_WCAG_PAIR_DISTANCE
-      [[5, 9]], // AA: distance 4, acceptable
-    );
-    expect(findDarkCounterpartIndex(5, cv)).toBe(9);
-  });
-
-  it('falls back to math inversion when all pairs too close', () => {
-    // Suppress the expected warning
-    const warn = console.warn;
-    console.warn = () => {};
-    try {
-      const cv = makeColorValueWithPairs(
-        [[5, 6]], // AAA: distance 1
-        [[5, 7]], // AA: distance 2
-      );
-      expect(findDarkCounterpartIndex(5, cv)).toBe(5); // 10 - 5
-    } finally {
-      console.warn = warn;
-    }
-  });
-
-  it('throws when no WCAG data exists', () => {
-    const cv = makeColorValueWithPairs([], []);
-    expect(() => findDarkCounterpartIndex(3, cv)).toThrow('No WCAG accessibility data');
-  });
-
-  it('searches higher for light indices (<=5)', () => {
-    const cv = makeColorValueWithPairs([
-      [2, 8],
-      [2, 3],
-    ]);
-    expect(findDarkCounterpartIndex(2, cv)).toBe(8);
-  });
-
-  it('searches lower for dark indices (>5)', () => {
-    const cv = makeColorValueWithPairs([
-      [8, 2],
-      [8, 7],
-    ]);
-    expect(findDarkCounterpartIndex(8, cv)).toBe(2);
-  });
-
-  it('respects MIN_WCAG_PAIR_DISTANCE constant', () => {
-    expect(MIN_WCAG_PAIR_DISTANCE).toBe(3);
-  });
-});
 
 describe('deriveGenerationRule', () => {
   it('assigns correct rules to semantic tokens', () => {
