+++ packages/shared/src/index.ts
@@ -1,3 +1,4 @@
 export * from './component-intelligence';
+export * from './scale-positions';
 export * from './types';
 export * from './version';
